requirements.txt
Pillow>=10.0.0
Flask>=3.0.0
PyYAML>=6.0
Babel>=2.12.0
rich>=13.0.0
inky[rpi]>=2.0.0
ffmpeg-python>=0.2.0
qrcode[pil]>=7.0.0
dacite>=1.8.0
RPi.GPIO>=0.7.0
